+++ condow_core/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.1] - 2021/09/21
+
+### ADDED
+
+* Range adapter for offset and length
+
+### CHANGED
+
+* Ranges in multi downloads must be `Clone`
 ## [0.9.0] - 2021/09/21
 
 ### ADDED